requirements.txt
alembic==1.6.5
aniso8601==9.0.1
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.6
cfgv==3.3.0
click==8.0.1
colorama==0.4.4
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.2.3
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
greenlet==1.1.0
identify==2.2.11
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
nodeenv==1.6.0
platformdirs==2.2.0
pre-commit==2.13.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
typing-extensions==3.10.0.0
virtualenv==20.6.0
Werkzeug==2.0.1
zipp==3.5.0
